// apps/documentum-core-frontend/src/types/api.ts
export interface Gamme {
  id: number;
  nom: string;
  description?: string | null;
  is_archived: boolean;
}
// ...
export interface VersionProjet {
  id: number;
  version_numero: string;    // ex: "1.0.0"
  date_lancement: string | null; // ISO datetime
  is_active: boolean;
  is_archived: boolean;
}

export interface Map {
  id: number;
  nom: string;
  projet: number;     // id du projet
  is_master: boolean; // racine « master »
}

export interface Project {
  id: number;
  nom: string;
  description: string;
  gamme: Gamme;                 // objet imbriqué (read_only)
  // En écriture, on envoie "gamme_id", mais ce champ n'est pas renvoyé en lecture
  // gamme_id?: number;         // write_only
  maps: Map[];                  // read_only
  versions: VersionProjet[];    // read_only
  version_numero?: string | null;
  date_creation: string;        // ISO datetime
  date_mise_a_jour: string;     // ISO datetime
  auteur: number | null;        // id de l'utilisateur
}